+++ lib/docker/DockerProgressReporter.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Atomist, Inc.
+ * Copyright © 2019 Atomist, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -36,6 +36,12 @@ export const DockerProgressTests: ProgressTest[] = [{
 }, {
     test: /Invoking goal hook: post/i,
     phase: "post-hook",
+}, {
+    test: /\/kaniko\/executor/i,
+    phase: "kaniko build",
+}, {
+    test: /pushed blob sha256/i,
+    phase: "kaniko push",
 }];
 
 /**